chore(repo): split angular e2e tests into core and extensions
