+++ e2e/angular-core/project.json
@@ -1,6 +1,6 @@
 {
-  "root": "e2e/angular",
-  "sourceRoot": "e2e/angular",
+  "root": "e2e/angular-core",
+  "sourceRoot": "e2e/angular-core",
   "projectType": "application",
   "targets": {
     "e2e": {
@@ -14,7 +14,7 @@
             "command": "yarn e2e-build-package-publish"
           },
           {
-            "command": "nx run-e2e-tests e2e-angular"
+            "command": "nx run-e2e-tests e2e-angular-core"
           }
         ],
         "parallel": false
@@ -23,11 +23,11 @@
     "run-e2e-tests": {
       "executor": "@nrwl/jest:jest",
       "options": {
-        "jestConfig": "e2e/angular/jest.config.js",
+        "jestConfig": "e2e/angular-core/jest.config.js",
         "passWithNoTests": true,
         "runInBand": true
       },
-      "outputs": ["coverage/e2e/angular"]
+      "outputs": ["coverage/e2e/angular-core"]
     }
   },
   "implicitDependencies": ["angular"]
