+++ e2e/angular-core/tsconfig.json
