+++ e2e/angular-core/tsconfig.spec.json
