+++ e2e/angular-extensions/project.json
@@ -1,0 +1,34 @@
+{
+  "root": "e2e/angular-extensions",
+  "sourceRoot": "e2e/angular-extensions",
+  "projectType": "application",
+  "targets": {
+    "e2e": {
+      "executor": "@nrwl/workspace:run-commands",
+      "options": {
+        "commands": [
+          {
+            "command": "yarn e2e-start-local-registry"
+          },
+          {
+            "command": "yarn e2e-build-package-publish"
+          },
+          {
+            "command": "nx run-e2e-tests e2e-angular-extensions"
+          }
+        ],
+        "parallel": false
+      }
+    },
+    "run-e2e-tests": {
+      "executor": "@nrwl/jest:jest",
+      "options": {
+        "jestConfig": "e2e/angular-extensions/jest.config.js",
+        "passWithNoTests": true,
+        "runInBand": true
+      },
+      "outputs": ["coverage/e2e/angular-extensions"]
+    }
+  },
+  "implicitDependencies": ["angular"]
+}
