+++ workspace.json
@@ -20,7 +20,8 @@
     "dep-graph-dep-graph-e2e": "dep-graph/dep-graph-e2e",
     "devkit": "packages/devkit",
     "docs": "docs",
-    "e2e-angular": "e2e/angular",
+    "e2e-angular-core": "e2e/angular-core",
+    "e2e-angular-extensions": "e2e/angular-extensions",
     "e2e-cli": "e2e/cli",
     "e2e-cypress": "e2e/cypress",
     "e2e-gatsby": "e2e/gatsby",
